refactor(all): Modify package structure and class name
Before
CarForwardCondition

After
RacingStrategy

자동차가 움직일 수 있는 조건을 제공하는 인터페이스를 선언했습니다. 경주 전략이라는 의미를 담아 더 직관적인 이름으로 변경했습니다.
