--- src/main/java/racingcar/domain/CarForwardCondition.java
@@ -1,5 +1,0 @@
-package racingcar.domain;
-
-public interface CarForwardCondition {
-    boolean determineCarToForward();
-}
