+++ src/main/java/racingcar/domain/racing/RacingStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.domain.racing;
+
+public interface RacingStrategy {
+    boolean determineCarToForward();
+}
