+++ src/test/java/racingcar/domain/CarTest.java
@@ -2,12 +2,12 @@ package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.lang.reflect.Field;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.common.exception.CarPositionOutOfRangeException;
+import racingcar.domain.car.Car;
 
 class CarTest {
